bot/requirements.txt
fastapi
uvicorn
aiogram
python-dotenv
requests
httpx
soundfile
supabase